Reworking UniFFI Rust calls
Changed how the JS scaffolding functions return the result of the Rust
call.  Instead of returning a value and filling out a
UniFFIRustCallStatus out param, they new return a single
UniFFIRustCallResult value, which contains the result code and either
the error or the return value.

The reason for the switch is async code.  To make that work, the
scaffolding methods need to return a Promise.  But it seems very weird
to mix a Promise with an out param.

There's a couple nice side benefits:
  - The return type / out param code can be simpler, since we always
    return the same type.
  - UniFFIRustCallResult is a dictionary, which can be auto-generated.
    No need to write C++ code for it.

Made some other changes as well:
  - Renamed `CALL_PANIC` to `CALL_INTERNAL_ERROR`.  We should never see
    Rust panic, since FF sets the panic=abort flag.  But we do have some
    internal errors that we want to handle.
  - Deleted the `Bindings.jsm` file which was no longer used.
  - Prefix the extra arguments in our scaffolding functions to avoid
    name conflicts with the rust arguments.