Adding some simple type checking to UniFFI generated code
Implemented type checks on:
  - Object constructors
  - Object method calls
  - Record constructors
  - Function calls

I added type checks to those functions, instead of the `lower()` method
was because we have access to the argument names and can generate nicer
error messages with that.  Also, for Records, it seems better to throw
an error at construction time rather than wait until we try to lower it.

Right now there are two checks:
  - All arguments are present
  - We can safely convert between a JS `number` type, which is floating
    point, and a 64-bit Rust integers.

We should be able to add more checks if we need them.

For async functions, which is all functions currently, I chose to make
them reject the promise rather than throw an exception synchronously.  That
felt like the nicer API to me, but I wasn't very confident about this
one.

Refactored the scaffolding function call code and moved it to a macro.
Now that it's more complex, I wanted to only have 1 implementation of
it.

Added a base `FfiConverter` class, so that types could share the base
`typeCheck()` method.