Better handling object pointers
We were using the JS `number` type for these, but that's problematic
since there's no guarantee that a pointer can be coerced to a number.
Instead, use the "private value API" of `JS::Value` to directly set the
pointer value.

In the short-term, this probably doesn't make any difference since all
current architectures use a of 48-bit address space at max, which can
safely be converted to a JS number.  But going forward, it seems better
to use an API designed to store pointers.

Note that we expose these values to the JS code, which is a bit weird.
However I think it's okay since a) it's only used for the `.ptr` field
in the generate code, which should be considered private and b) we treat
it as an opaque value and never modify it.

Updated the scaffolding rendering code to handle this.  I tried to
clarify how arguments/return values get transformed as we make
scaffolding calls.  In particular the `FFIType.type_name()` method seems
misnamed since we can use up to 3 different types to handle arguments.